Fix unspecified List deserialization
This code:

```
import jsons
from typing import List

jsons.load(["Hello", "World"], List)
```

would fail with a: TypeError: 'NoneType' object is not callable.

This is because the bare List type contains an __args__[0] which is a TypeVar, which is not callable when deserializing.

This pull request fixed that by checking if the __args__[0] is actually usable and not just a bare TypeVar.

This error occurs specifically when explicitly casting to a List, but not specifying the type of the list items.

Two tests are added to highlight this.