Correctly handle nested objects with verbose type hints
The previous implementation of `new_hints` (for recursing on type hints
while deserializing) used `str.replace`.  In the case of nested objects
with identical property names, that would give an incorrect result.  We
actually want to remove the `cls_key` from all keys it is a prefix of,
but not touch non-prefix occurrences in the string.  This lets us
correctly recurse:
```
/inner/inner
  |
  v
/inner
  |
  v
/
```